webapi/webapi-messaging-sysapps-tests/messaging/SmsMessage_basic.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Jianfeng.Xu <jianfengx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>Messaging Test: SmsMessage_Basic</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/messaging/#idl-def-SmsMessage">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/support.js"></script>
<div id="log"></div>
<script>
var testNumber = "5556";
var filter = {
  type: "sms",
};
var filterOption = {
  sortBy: "date",
  sortBy: "descending",
  limit: 3
};
var t1 = async_test("Check if the readonly attribute messageID of SmsMessage exists and type of string", {timeout: 2000});
var t2 = async_test("Check if the readonly attribute type of SmsMessage exists and type of string", {timeout: 2000});
var t3 = async_test("Check if the readonly attribute serviceID of SmsMessage exists and type of string", {timeout: 2000});
var t4 = async_test("Check if the readonly attribute from of SmsMessage exists and type of string", {timeout: 2000});
var t5 = async_test("Check if the readonly attribute timestamp of SmsMessage exists and type of string", {timeout: 2000});
var t6 = async_test("Check if the readonly attribute read of SmsMessage exists and type of boolean", {timeout: 2000});
var t7 = async_test("Check if the readonly attribute to of SmsMessage exists and type of string", {timeout: 2000});
var t8 = async_test("Check if the readonly attribute body of SmsMessage exists and type of string", {timeout: 2000});
var t9 = async_test("Check if the readonly attribute state of SmsMessage exists and type of string", {timeout: 2000});
var t10 = async_test("Check if the readonly attribute deliveryStatus of SmsMessage exists and type of string", {timeout: 2000});
var t11 = async_test("Check if the readonly attribute deliveryTimestamp of SmsMessage exists and type of Date", {timeout: 2000});
var t12 = async_test("Check if the readonly attribute messageClass of SmsMessage exists and type of  string", {timeout: 2000});


t1.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t1.step(function () {
        assert_true(msg.element.length > 0, "no sms message");
        var smsmessage = msg.element[0];
        assert_true("messageID" in smsmessage, "The messageID attribute of SmsMessage exists");
        var dc = smsmessage.messageID;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type messageID of SmsMessage should be string");
        smsmessage.messageID = dc + "test";
        assert_equals(smsmessage.messageID, dc, "The value of messageID");
      });
      t1.done();
    },
    function (error) {
      t1.step(function () {
        assert_unreached(error.message);
      });
      t1.done();
    }
  );
});

t2.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t2.step(function () {
        assert_true(msg.element.length > 0, "no sms message");
        var smsmessage = msg.element[0];
        assert_true("type" in smsmessage, "The type attribute of SmsMessage exists");
        var dc = smsmessage.type;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of SmsMessage.type should be string");
        smsmessage.type = dc + "test";
        assert_equals(smsmessage.type, dc, "The value of type");
      });
      t2.done();
    },
    function (error) {
      t2.step(function () {
        assert_unreached(error.message);
      });
      t2.done();
    }
  );
});

t3.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t3.step(function () {
        assert_true(msg.element.length > 0, "no sms message");
        var smsmessage = msg.element[0];
        assert_true("serviceID" in smsmessage, "The serviceID attribute of SmsMessage exists");
        var dc = smsmessage.serviceID;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "type of SmsMessage.serviceID should be string");
        smsmessage.serviceID = dc + "test";
        assert_equals(smsmessage.serviceID, dc, "The value of serviceID");
      });
      t3.done();
    },
    function (error) {
      t3.step(function () {
        assert_unreached(error.message);
      });
      t3.done();
    }
  );
});

t4.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t4.step(function () {
        assert_true(msg.element.length > 0, "no sms message");
        var smsmessage = msg.element[0];
        assert_true("from" in smsmessage, "The from attribute of SmsMessage exists");
        var dc = smsmessage.from;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of SmsMessage.from should be string");
        smsmessage.from = dc + "test";
        assert_equals(smsmessage.from, dc, "The value of from");
      });
      t4.done();
    },
    function (error) {
      t4.step(function () {
        assert_unreached(error.message);
      });
      t4.done();
    }
  );
});

t5.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t5.step(function () {
        assert_true(msg.element.length > 0, "no sms message");
        var smsmessage = msg.element[0];
        assert_true("timestamp" in smsmessage, "The timestamp attribute of SmsMessage exists");
        var dc = smsmessage.timestamp;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of SmsMessage.timestamp should be string");
        smsmessage.timestamp = dc + "test";
        assert_equals(smsmessage.serviceID, dc, "The value of timestamp");
      });
      t5.done();
    },
    function (error) {
      t5.step(function () {
        assert_unreached(error.message);
      });
      t5.done();
    }
  );
});

t6.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t6.step(function () {
        assert_true(msg.element.length > 0, "no sms message");
        var smsmessage = msg.element[0];
        assert_true("read" in smsmessage, "The read attribute of SmsMessage exists");
        var dc = smsmessage.read;
        assert_equals(Object.prototype.toString.call(dc), "[object boolean]", "The type of SmsMessage.read should be boolean");
        if(dc) {
          smsMessage.read = false;
        } else {
          smsMessage.read = true;
        }
        assert_equals(smsmessage.read, dc, "The value of read");
      });
      t6.done();
    },
    function (error) {
      t6.step(function () {
        assert_unreached(error.message);
      });
      t6.done();
    }
  );
});

t7.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t7.step(function () {
        assert_true(msg.element.length > 0, "no sms message");
        var smsmessage = msg.element[0];
        assert_true("to" in smsmessage, "The to attribute of SmsMessage exists");
        var dc = smsmessage.to;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of SmsMessage.to should be string");
        smsmessage.to = dc + "test";
        assert_equals(smsmessage.to, dc, "The value of to");
      });
      t7.done();
    },
    function (error) {
      t7.step(function () {
        assert_unreached(error.message);
      });
      t7.done();
    }
  );
});

t8.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t8.step(function () {
        assert_true(msg.element.length > 0, "no sms message");
        var smsmessage = msg.element[0];
        assert_true("body" in smsmessage, "The body attribute of SmsMessage exists");
        var dc = smsmessage.body;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of SmsMessage.body should be string");
        smsmessage.body = dc + "test";
        assert_equals(smsmessage.body, dc, "The value of body");
      });
      t8.done();
    },
    function (error) {
      t8.step(function () {
        assert_unreached(error.message);
      });
      t8.done();
    }
  );
});

t9.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t9.step(function () {
        assert_true(msg.element.length > 0, "no sms message");
        var smsmessage = msg.element[0];
        assert_true("state" in smsmessage, "The state attribute of SmsMessage exists");
        var dc = smsmessage.state;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of SmsMessage.state should be string");
        smsmessage.state = dc + "test";
        assert_equals(smsmessage.state, "test", "The value of state");
      });
      t9.done();
    },
    function (error) {
      t9.step(function () {
        assert_unreached(error.message);
      });
      t9.done();
    }
  );
});

t10.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t10.step(function () {
        assert_true(msg.element.length > 0, "no sms message");
        var smsmessage = msg.element[0];
        assert_true("deliveryStatus" in smsmessage, "The serviceID attribute of SmsMessage exists");
        var dc = smsmessage.deliveryStatus;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of SmsMessage.deliveryStatus should be string");
        smsmessage.deliveryStatus = dc + "test";
        assert_equals(smsmessage.deliveryStatus, dc, "The value of deliveryStatus");
      });
      t10.done();
    },
    function (error) {
      t10.step(function () {
        assert_unreached(error.message);
      });
      t10.done();
    }
  );
});

t11.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t11.step(function () {
        assert_true(msg.element.length > 0, "no sms message");
        var smsmessage = msg.element[0];
        assert_true("deliveryTimestamp" in smsmessage, "The deliveryTimestamp attribute of SmsMessage exists");
        var dc = smsmessage.deliveryTimestamp;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of SmsMessage.deliveryTimestamp should be string");
        smsmessage.deliveryTimestamp = dc + "test";
        assert_equals(smsmessage.deliveryTimestamp, dc, "The value of deliveryTimestamp");
      });
      t11.done();
    },
    function (error) {
      t11.step(function () {
        assert_unreached(error.message);
      });
      t11.done();
    }
  );
});

t12.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t12.step(function () {
        assert_true(msg.element.length > 0, "no sms message");
        var smsmessage = msg.element[0];
        assert_true("messageClass" in smsmessage, "The messageClass attribute of SmsMessage exists");
        var dc = smsmessage.messageClass;
        assert_equals(Object.prototype.toString.call(dc), "[object String]", "The type of SmsMessage.messageClass should be string");
        smsmessage.messageClass = dc + "test";
        assert_equals(smsmessage.messageClass, dc, "The value of messageClass");
      });
      t12.done();
    },
    function (error) {
      t12.step(function () {
        assert_unreached(error.message);
      });
      t12.done();
    }
  );
});

</script>
